Refactored paths
