+++ src/main/java/org/wso2/carbon/module/swiftiso20022/mt/parsers/MT940Parser.java
@@ -19,7 +19,7 @@
 package org.wso2.carbon.module.swiftiso20022.mt.parsers;
 
 import org.wso2.carbon.module.swiftiso20022.exceptions.MTMessageParsingException;
-import org.wso2.carbon.module.swiftiso20022.mt.models.mtmessages.MT940Message;
+import org.wso2.carbon.module.swiftiso20022.mt.models.messages.MT940Message;
 import org.wso2.carbon.module.swiftiso20022.utils.MTParserUtils;
 
 import java.util.Map;
