+++ src/test/java/org/wso2/carbon/module/swiftiso20022/mtmessageparser/MTParserTests.java
@@ -25,7 +25,7 @@ import org.testng.annotations.Test;
 import org.wso2.carbon.module.swiftiso20022.exceptions.MTMessageParsingException;
 import org.wso2.carbon.module.swiftiso20022.mt.models.blocks.ApplicationHeaderBlock;
 import org.wso2.carbon.module.swiftiso20022.mt.models.blocks.BasicHeaderBlock;
-import org.wso2.carbon.module.swiftiso20022.mt.models.mtmessages.MT940Message;
+import org.wso2.carbon.module.swiftiso20022.mt.models.messages.MT940Message;
 import org.wso2.carbon.module.swiftiso20022.mt.parsers.MT940Parser;
 import org.wso2.carbon.module.swiftiso20022.mt.parsers.MTParser;
 import org.wso2.carbon.module.swiftiso20022.utils.MTParserConstants;
